+++ src/main/java/com/fitlogtimer/dto/ExerciseSetOutDTO.java
@@ -1,11 +1,11 @@
 package com.fitlogtimer.dto;
 
 public record ExerciseSetOutDTO(
-        Long id,
+        int id,
         int exercise_id,
         double weight,
         int repNumber,
         boolean isMax,
         String comment,
-        Long session_id) {
+        int session_id) {
 }
